Very minor reorg
Move transform stage into its own file, cargo fmt.
